requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5
streamlit>=1.28.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
requests>=2.31.0
PyMuPDF>=1.22.5
Pillow>=9.5.0
mistralai>=0.0.7
python-dotenv>=1.0.0
pydantic>=2.5.2